move some functions out of `chat.py`
